[YITEAM-84] 프로필 API 수정
- 이미지 업로드 API 분리
