+++ src/main/java/com/yapp/ios1/controller/S3Controller.java
@@ -1,16 +1,25 @@
 package com.yapp.ios1.controller;
 
+import com.yapp.ios1.dto.ResponseDto;
 import com.yapp.ios1.service.S3Service;
+import com.yapp.ios1.utils.auth.Auth;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.List;
+
+import static com.yapp.ios1.common.ResponseMessage.NOT_EXIST_IMAGE;
+import static com.yapp.ios1.common.ResponseMessage.UPLOAD_IMAGE_SUCCESS;
 
 /**
  * created by ayoung 2021/03/29
  */
+@Api(tags = "Image Upload")
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/api/v2")
@@ -19,10 +28,22 @@ public class S3Controller {
     private final S3Service s3Service;
 
     /**
-     * S3 파일 업로드를 위한 임시 코드 (삭제 예정)
+     * @param imageList 이미지 리스트
      */
-    @PostMapping("/upload")
-    public List<String> upload(@RequestParam("file") MultipartFile[] multipartFile) throws IOException {
-        return s3Service.upload(multipartFile);
+    @ApiOperation(
+            value = "이미지 업로드",
+            notes = "이미지 url 배열 리턴"
+    )
+    @Auth
+    @PostMapping("/images")
+    public ResponseEntity<ResponseDto> registerBucketImageList(@RequestParam(value = "image") MultipartFile[] imageList) throws IOException {
+        if (imageList == null) {
+            ResponseEntity.ok()
+                    .body(ResponseDto.of(HttpStatus.BAD_REQUEST, NOT_EXIST_IMAGE));
+        }
+
+        return ResponseEntity.ok()
+                .body(ResponseDto.of(HttpStatus.OK, UPLOAD_IMAGE_SUCCESS, s3Service.upload(imageList)));
     }
+
 }
