+++ src/main/java/com/yapp/ios1/controller/UserController.java
@@ -20,9 +20,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
 import java.sql.SQLException;
 import java.util.Optional;
 
@@ -122,15 +120,13 @@ public class UserController {
     }
 
     @ApiOperation(
-            value = "프로필 수정",
-            notes = "포스트맨에서 테스트 가능"
+            value = "프로필 수정"
     )
     @Auth
     @PutMapping("")
-    public ResponseEntity<ResponseDto> updateProfile(@RequestPart(value = "profileImage", required = false) MultipartFile profileImage,
-                                                     @RequestPart ProfileDto profile) throws IOException {
+    public ResponseEntity<ResponseDto> updateProfile(@RequestBody ProfileDto profile) {
         Long userId = UserContext.getCurrentUserId();
-        userService.updateProfile(profile, profileImage, userId);
+        userService.updateProfile(profile, userId);
         return ResponseEntity.ok().body(ResponseDto.of(HttpStatus.OK, UPDATE_PROFILE_SUCCESS));
     }
