+++ src/main/java/com/yapp/ios1/service/UserService.java
@@ -13,13 +13,10 @@ import com.yapp.ios1.mapper.FollowMapper;
 import com.yapp.ios1.mapper.UserMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
 import java.sql.SQLException;
 import java.util.List;
 import java.util.Optional;
@@ -35,13 +32,9 @@ import static com.yapp.ios1.common.ResponseMessage.*;
 public class UserService {
 
     private final BucketService bucketService;
-    private final S3Service s3Service;
     private final UserMapper userMapper;
     private final PasswordEncoder passwordEncoder;
     private final FollowMapper followMapper;
-
-    @Value("${buok.s3.dir.profile}")
-    private String profileDir;
 
     /**
      * 이메일 존재하는지 확인
@@ -124,12 +117,8 @@ public class UserService {
 
     // 프로필 업데이트
     @Transactional
-    public void updateProfile(ProfileDto profileDto, MultipartFile profileImage, Long userId) throws IOException {
-        profileDto.setUserId(userId);
-        if (profileImage != null) {
-            profileDto.setProfileUrl(s3Service.upload(profileImage, profileDir));
-        }
-        userMapper.updateProfile(profileDto);
+    public void updateProfile(ProfileDto profileDto, Long userId) {
+        userMapper.updateProfile(profileDto, userId);
     }
 
     @Transactional(readOnly = true)
